Create a seperate tour and user file
