+++ routes/tourRoutes.js
@@ -1,0 +1,82 @@
+const express = require('express');
+const fs = require('fs');
+
+const router = express.Router();
+
+const tours = JSON.parse(
+  fs.readFileSync(`${__dirname}/../dev-data/data/tours-simple.json`)
+);
+
+const getAllTours = (req, res) => {
+  console.log(req.requestTime);
+  res.status(200).json({
+    status: 'success',
+    requestedAt: req.requestTime,
+    results: tours.length,
+    data: { tours },
+  });
+};
+
+const getTour = (req, res) => {
+  console.log(req.params);
+  const id = +req.params.id;
+  const tour = tours.find((el) => el.id === id);
+
+  if (!tour) {
+    return res.status(404).json({ message: 'Invalid id' });
+  }
+
+  res.status(200).json({
+    status: 'success',
+    data: { tour },
+  });
+};
+
+const createTour = (req, res) => {
+  console.log(req.body);
+
+  const newId = tours[tours.length - 1].id + 1;
+  const newTour = Object.assign({ id: newId }, req.body);
+
+  tours.push(newTour);
+  fs.writeFile(
+    `${__dirname}/dev-data/data/tours-simple.json`,
+    JSON.stringify(tours),
+    (err) => {
+      res.status(201).json({
+        status: 'success',
+        data: {
+          tour: newTour,
+        },
+      });
+    }
+  );
+};
+
+const updateTour = (req, res) => {
+  if (+req.params.id > tours.length) {
+    return res.status(404).json({ message: 'Invalid id' });
+  }
+
+  res.status(200).json({
+    status: 'success',
+    data: { tour: 'update tour' },
+  });
+};
+
+const deleteTour = (req, res) => {
+  if (+req.params.id > tours.length) {
+    return res.status(404).json({ message: 'Invalid id' });
+  }
+
+  res.status(204).json({
+    status: 'success',
+    data: null,
+  });
+};
+
+router.route('/').get(getAllTours).post(createTour);
+
+router.route('/:id').get(getTour).patch(updateTour).delete(deleteTour);
+
+module.exports = router;
